Added controllers, services and DAOs

+++ core/src/main/java/com/javanostra/spring/core/entities/ChatMemberId.java
@@ -3,6 +3,6 @@ package com.javanostra.spring.core.entities;
 import java.io.Serializable;
 
 public class ChatMemberId implements Serializable {
-    private Integer chat;
-    private Integer user;
+    private Long chat;
+    private Long user;
 }
